+++ src/main/java/com/chatop/repository/RentalRepository.java
@@ -3,5 +3,5 @@ package com.chatop.repository;
 import com.chatop.model.Rental;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RentalRepository extends JpaRepository<Rental, Integer> {
+public interface RentalRepository extends JpaRepository<Rental, Long> {
 }
